requirements.txt
paho-mqtt==2.1.0
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.2.0
tensorflow==2.13.0
pymongo==4.5.0
plotly==5.18.0
dash==2.14.0
python-dotenv==1.0.0